v3.1 update saved_entry only if --make-default specified
This update is in response to rhbz#1111790

This change to grubby.c implements updating saved_entry in
/boot/grub2/grubenv only if the --make-default is specified on
grubby's command line.  Another update has changed new-kernel-pkg so
that --make-default is not specified unless UPDATEDEWFAULT=yes in
/etc/sysconfig/kernel.

Both defaultKernel (--set-default=) and defaultIndex (--set-default-index=)
are also tested so that the tests run properly.