v1.3 if GRUBBY_ENABLE_DEBUG=yes then add --enable-debug
In the new-kernel-pkg shell script add --enable-debug to the
grubby command line if GRUBBY_ENABLE_DEBUG=yes.  This will also
set verbose="-v" fopr more info from new-kernel-pkg.

This will cause a large amount of debugging information to be
printed when grubby is run.  This is not enabled during
tests.