v3.2 add support for btrfs when grub2 is bootloader
This patch is in to response to RHBZ#1094489.

This patch adds getSubvolPrefix() for handling a btrfs subvol
prefix on the filenames for grub2.  If get_Root_Specifier() results
in a NULL rootspec, then getSubvolPrefix() is executed to extract any
btrfs subvol prefix.  With this modification, booting off
/boot on a btrfs subvol is supported.

To determine if booting is to be performed off a btrfs volume
or subvolume, the lines of the entry are scanned to detect
if "insmod btrfs" is present.  If it is, then btrfs is assumed.

If the kernel filename includes only one slash, then there is
no btrfs prefix.  Otherwise, the first part defined by "/../"
will be the btrfs prefix if booting btrfs.  Unless, of course,
there are 2 slashes in the filename but bootPrefix is zero length
in which case we are booting off a btrfs volume!

Besides the kernel, the subvol prefix must also be on the
initrd filename.  The subvol prefix for initrd is based on
the entry's kernel's filename.  Note that the lines of
the entry must be scanned to determine if we are booting on
btrfs.

Code was added to addLine(tmpl() so that the subvol prefix not added
for an initrd since updateInitrd() does that already.

This patch includes a whole lot of debugging code in support of
implementing btrfs support.  The debug output will only be produced
if --enable-debug is a grubby command line argument.  Thus, this
patch depends on the patch(es) implementing a run-time debug option.